fix: addressed code review changes, and missed bug
fixed an infinite render issue, as well as a missed click back link and simplified the ux. created a reusable error component for the 4xx pages
