+++ frontend/src/app/components/containers/pages/verify-access.tsx
@@ -1,14 +1,11 @@
 import { FC, useEffect, useMemo, useState } from "react";
 import { Link, useLocation } from "react-router-dom";
-import logoLg from "../../../../assets/images/branding/BCgov-lg.png";
-import { Footer } from "../layout";
+import { Footer, Header } from "../layout";
 import { TbFaceId, TbFaceIdError, TbLockAccess } from "react-icons/tb";
 import { validate as uuidValidate } from "uuid";
 import { useAppDispatch, useAppSelector } from "../../../hooks/hooks";
 import { validateComsAccess } from "../../../store/reducers/app";
-import config from "../../../../config";
-import EnvironmentBanner from "../layout/environment-banner";
-import { IconType } from "react-icons/lib";
+import { ErrorMessage } from "./error-message";
 
 type Props = {};
 
@@ -20,18 +17,14 @@ const useQuery = () => {
 export const VerifyAccess: FC<Props> = () => {
   const dispatch = useAppDispatch();
   const query = useQuery();
-
   const [isValid, setIsValid] = useState<boolean | undefined>();
-  const [wasRun, setWasRun] = useState(false);
 
-  const environmentName = config.ENVIRONMENT_NAME || "production";
   const isLoading = useAppSelector((state) => state?.app?.loading?.isLoading);
+  const token = query.get("token");
 
   useEffect(() => {
-    const token = query.get("token");
-
-    if (!isLoading && !wasRun) {
-      if (token && uuidValidate(token) && !wasRun) {
+    if (!isLoading) {
+      if (token && uuidValidate(token)) {
         dispatch(validateComsAccess(token)).then((res) => {
           const { status } = res;
           if (status !== "success") {
@@ -39,42 +32,46 @@ export const VerifyAccess: FC<Props> = () => {
           }
           if (status === "success") {
             setIsValid(true);
-            setWasRun(true);
-            window.location.href = "/complaints";
+
+            setTimeout(() => {
+              window.location.href = "/complaints";
+            }, 3000);
           }
         });
       }
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [query, isLoading, wasRun]);
+  }, [token]);
 
   const renderResult = () => {
     if (isValid === undefined) {
       return (
-        <Message
+        <ErrorMessage
           icon={TbLockAccess}
           header="Verifying access"
         >
           Your access is being verified. You will be redirected to the application once verified. <br />
           If you have any problems, please contact the Compliance and Enforcement Digital Service team at{" "}
           <a href="mailto:CEDS@gov.bc.ca">CEDS@gov.bc.ca</a>
-        </Message>
+        </ErrorMessage>
       );
     } else if (isValid) {
       return (
-        <Message
+        <ErrorMessage
           icon={TbFaceId}
           header="Access granted"
         >
           Your access has been verified. You will be redirected to the application shortly. <br />
-          If you are not redirected in 5 seconds click the link to continue: <Link to="/" /> <br />
+          If you are not redirected in 5 seconds click the link to continue:{" "}
+          <Link to="/complaints">NatComplaints - List View</Link> <br />
           If you have any problems, please contact the Compliance and Enforcement Digital Service team at{" "}
           <a href="mailto:CEDS@gov.bc.ca">CEDS@gov.bc.ca</a>
-        </Message>
+        </ErrorMessage>
       );
     } else {
       return (
-        <Message
+        <ErrorMessage
           icon={TbFaceIdError}
           header="Access denied"
         >
@@ -82,75 +79,18 @@ export const VerifyAccess: FC<Props> = () => {
           Your access link has expired or is incorrect. <br />
           Please contact the Compliance and Enforcement Digital Services team at{" "}
           <a href="mailto:CEDS@gov.bc.ca">CEDS@gov.bc.ca</a> to request a new link to get access.
-        </Message>
+        </ErrorMessage>
       );
     }
   };
 
   return (
-    <>
-      {environmentName !== "production" && <EnvironmentBanner environmentName={environmentName} />}
-      <div className="comp-app-container fixed-header">
-        {/* <!-- --> */}
+    <div className="comp-app-container">
+      <Header />
 
-        <div className="comp-header">
-          <Link
-            className="comp-header-logo"
-            to="/"
-          >
-            <picture>
-              <source srcSet={logoLg}></source>
-              <img
-                src={logoLg}
-                alt={"Government of British Columbia"}
-              />
-            </picture>
-            NatComplaints
-          </Link>
+      <div className="error-container">{renderResult()}</div>
 
-          <div className="comp-header-content">
-            <div className="comp-header-left">{/* <!-- future left hand content --> */}</div>
-            <div className="comp-header-right">
-              <div className="header-btn-lg pr-0">
-                <div className="widget-content p-0">
-                  <div className="widget-content-wrapper">
-                    <div className="widget-content-left">{/* <!-- search --> */}</div>
-                    <div className="widget-content-left"></div>
-                    <div className="widget-content-right">
-                      {/* <!-- --> */}
-
-                      {/* <!-- --> */}
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        {/* <!-- --> */}
-
-        <div className="error-container">{renderResult()}</div>
-
-        <Footer />
-      </div>
-    </>
-  );
-};
-
-type props = {
-  icon: IconType;
-  header: string;
-  children: React.ReactNode;
-};
-
-const Message: FC<props> = ({ icon, header, children }) => {
-  const Icon = icon;
-
-  return (
-    <div className="message">
-      <Icon />
-      <h1 className="comp-padding-top-25">{header}</h1>
-      {children}
+      <Footer />
     </div>
   );
 };
